Update renovate.json: detect patch updates
On this repository, the conventional method which can be seen at other roles' repositories to have Renovate find updates would cause every version value on main.yml to be updated to the latest major version. Since this is not expected here, this commit edits renovate.json and main.yml for the specific need to find patch updates only, avoiding major or minor ones.

Referred resources:
- https://github.com/renovatebot/renovate/discussions/25248
- https://docs.renovatebot.com/configuration-options/

By setting matchUpdateTypes=major,minor with enabled=false, Renovate is configured to detect patch updates only. `separateMinorPatch` is also necessary because by default Removate groups patch and minor updates into a single PR. See: https://docs.renovatebot.com/configuration-options/#separateminorpatch

As noted on main.yml, adding a variable or editing one manually is required for major or minor updates for safety.

`*Template` options are specified in a `customManagers` object as the values are not captured with capture groups in `matchString`. As the value for `matchStrings` is not limited to lines for a version setting, `versioningTemplate` limits what to be captured instead. This template should be able to be removed safely by improving the `matchString`'s value.

This works at least, but obviously there should be room for improvement.

Signed-off-by: Suguru Hirahara <acioustick@noreply.codeberg.org>